apps/Backend/requirements.txt
langchain
python-dotenv
langchain_community
langserve
fastapi
uvicorn
langchain-ollama
sse_starlette
pydantic
gunicorn
